alertas/requirements.txt
pandas
openpyxl
PyYAML
python-dotenv
pytest
